move types to centralized place and add prettier and eslint for them
