+++ .github/workflows/jira_link_to_pr.yml
@@ -5,19 +5,19 @@
 #   follow the pattern: <username>:<api token from jira>
 name: jira-description-action
 on:
-  pull_request:
-    types: [opened, reopened]
+    pull_request:
+        types: [ opened, reopened ]
 jobs:
-  add-jira-description:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: cakeinpanic/jira-description-action@v0.9.0
-        name: jira-description-action
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          jira-token: ${{ secrets.JIRA_TOKEN }}
-          jira-base-url: https://hackolade.atlassian.net
-          skip-branches: '^(develop|main|master)$' #optional 
-          jira-project-key: 'HCK' #optional
-          use: 'both'
-          fail-when-jira-issue-not-found: false
+    add-jira-description:
+        runs-on: ubuntu-latest
+        steps:
+            -   uses: cakeinpanic/jira-description-action@v0.9.0
+                name: jira-description-action
+                with:
+                    github-token: ${{ secrets.GITHUB_TOKEN }}
+                    jira-token: ${{ secrets.JIRA_TOKEN }}
+                    jira-base-url: https://hackolade.atlassian.net
+                    skip-branches: '^(develop|main|master)$' #optional
+                    jira-project-key: 'HCK' #optional
+                    use: 'both'
+                    fail-when-jira-issue-not-found: false
