+++ forward_engineering/mappers/directiveUsageStatements.js
@@ -1,5 +1,5 @@
 /**
- * @import { DirectivePropertyData, IdToNameMap } from "../types/types"
+ * @import {DirectivePropertyData, IdToNameMap} from "../../shared/types/types"
  */
 
 const { joinInlineStatements } = require('../helpers/feStatementJoinHelper');
@@ -8,7 +8,7 @@ const { isUUID } = require('../helpers/isUUID');
 /**
  * Gets the directives usage statement by mapping directives to strings and joining them.
  *
- * @param {Object} params
+ * @param {object} params
  * @param {DirectivePropertyData[]} params.directives - Array of directive definitions
  * @param {IdToNameMap} [params.definitionsIdToNameMap] - The definitions id to name map.
  * @returns {string} - The joined directive statements
@@ -22,7 +22,8 @@ function getDirectivesUsageStatement({ directives = [], definitionsIdToNameMap =
 
 /**
  * Maps a single directive to its string representation with name and arguments.
- * @param {Object} params - The parameters object
+ *
+ * @param {object} params - The parameters object
  * @param {DirectivePropertyData} params.directive - The directive to map
  * @param {IdToNameMap} [params.definitionsIdToNameMap] - The definitions id to name map.
  * @returns {string} - The directive statement or empty string if invalid
@@ -45,9 +46,9 @@ function mapDirective({ directive, definitionsIdToNameMap }) {
 }
 
 /**
- * Maps a directive's raw arguments to a formatted string.
- * Handles adding parentheses and cleaning up newlines.
- * @param {Object} params
+ * Maps a directive's raw arguments to a formatted string. Handles adding parentheses and cleaning up newlines.
+ *
+ * @param {object} params
  * @param {DirectivePropertyData} params.directive - The directive containing the arguments
  * @returns {string} - The formatted arguments string or empty string if no valid arguments
  */
@@ -67,7 +68,8 @@ function mapDirectiveRawArguments({ directive }) {
 
 /**
  * Resolves and formats a directive name, ensuring proper @ prefix.
- * @param {Object} params
+ *
+ * @param {object} params
  * @param {string} params.directiveName - The raw directive name or ID
  * @param {IdToNameMap} [params.definitionsIdToNameMap] - The definitions id to name map.
  * @returns {string} - The formatted directive name or empty string if invalid
@@ -85,7 +87,8 @@ function getDirectiveName({ directiveName, definitionsIdToNameMap }) {
 
 /**
  * Formats a raw directive string by trimming whitespace and normalizing newlines.
- * @param {Object} params
+ *
+ * @param {object} params
  * @param {string} params.rawDirective - The raw directive string to format
  * @returns {string} - The formatted directive string or empty string if invalid
  */
