+++ forward_engineering/mappers/enums.js
@@ -1,19 +1,19 @@
 /**
- * @import { FEStatement, DirectivePropertyData, IdToNameMap } from "../types/types"
+ * @import {FEStatement, DirectivePropertyData, IdToNameMap} from "../../shared/types/types"
  */
 
 const { joinInlineStatements } = require('../helpers/feStatementJoinHelper');
 const { getDirectivesUsageStatement } = require('./directiveUsageStatements');
 
 /**
- * @typedef {Object} EnumValue
+ * @typedef {object} EnumValue
  * @property {string} value - The name of the enum value.
  * @property {string} description - The description of the enum value.
  * @property {DirectivePropertyData[]} typeDirectives - The directives of the enum value.
  */
 
 /**
- * @typedef {Object} EnumDefinition
+ * @typedef {object} EnumDefinition
  * @property {string} description - The description of the enum.
  * @property {boolean} isActivated - Indicates if the enum is activated.
  * @property {DirectivePropertyData[]} typeDirectives - The directives of the enum.
@@ -21,13 +21,13 @@ const { getDirectivesUsageStatement } = require('./directiveUsageStatements');
  */
 
 /**
- * @typedef {Object.<string, EnumDefinition>} EnumDefinitions
+ * @typedef {Record<string, EnumDefinition>} EnumDefinitions
  */
 
 /**
  * Gets the enums as an array of FEStatements.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {EnumDefinitions} param0.enumsDefinitions - The enums object.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
  * @returns {FEStatement[]}
@@ -41,7 +41,7 @@ function getEnums({ enumsDefinitions, definitionsIdToNameMap }) {
 /**
  * Maps a enum to an FEStatement.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {string} param0.name - The name of the enum.
  * @param {EnumDefinition} param0.enumDefinition - The enum definition object.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
@@ -65,7 +65,7 @@ function mapEnum({ name, enumDefinition, definitionsIdToNameMap }) {
 /**
  * Maps the enum values to an array of FEStatement.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {EnumValue[]} param0.enumValues - The enum values.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
  * @returns {FEStatement[]}
