+++ forward_engineering/mappers/objectLikeType.js
@@ -1,5 +1,5 @@
 /**
- * @import { FEStatement, DirectivePropertyData, ObjectLikeTypeDefinition, ObjectLikeTypeDefinitions, IdToNameMap, ImplementsInterface } from "../types/types"
+ * @import {FEStatement, ObjectLikeTypeDefinition, ObjectLikeTypeDefinitions, IdToNameMap} from "../../shared/types/types"
  */
 
 const { joinInlineStatements } = require('../helpers/feStatementJoinHelper');
@@ -9,7 +9,7 @@ const { getImplementsInterfacesStatement } = require('./implementsInterfaces');
 /**
  * Gets the object-like types from the model definitions.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {ObjectLikeTypeDefinitions} param0.objectTypes - The object-like types to get.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
  * @param {string} param0.typeKeyword - The type keyword ("type", "interface", "input").
@@ -25,7 +25,7 @@ function getObjectLikeTypes({ objectTypes, definitionsIdToNameMap, typeKeyword, 
 /**
  * Maps an object-like type to an FEStatement.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {string} param0.name - The name of the object.
  * @param {ObjectLikeTypeDefinition} param0.objectType - The object-like type definition object.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
