+++ forward_engineering/mappers/rootTypes.js
@@ -1,5 +1,5 @@
 /**
- * @import { FEStatement, IdToNameMap, RootTypeNamesParameter, EntityIdToJsonSchemaMap, EntityIdToPropertiesMap, ContainerDetails } from "../types/types"
+ * @import {FEStatement, IdToNameMap, RootTypeNamesParameter, EntityIdToJsonSchemaMap, EntityIdToPropertiesMap, ContainerDetails} from "../../shared/types/types"
  */
 
 const { QUERY_ROOT_TYPE, MUTATION_ROOT_TYPE, SUBSCRIPTION_ROOT_TYPE } = require('../constants/feScriptConstants');
@@ -8,14 +8,14 @@ const { getDirectivesUsageStatement } = require('./directiveUsageStatements');
 const { getRootTypeFields } = require('./fields');
 
 /**
- * Gets the root schema statement.
- * If all root types are empty, return null, to trigger validation error that we are referencing non-existing root types.
- * If the root type is not present in the root types, do not include it in the root schema statement.
+ * Gets the root schema statement. If all root types are empty, return null, to trigger validation error that we are
+ * referencing non-existing root types. If the root type is not present in the root types, do not include it in the root
+ * schema statement.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {RootTypeNamesParameter} param0.rootTypeNames - The root type names.
  * @param {FEStatement[]} param0.rootTypeStatements - The root types.
- * @param {ContainerDetails|undefined} param0.containerProperties - The container properties.
+ * @param {ContainerDetails | undefined} param0.containerProperties - The container properties.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
  * @returns {FEStatement | null} - The root schema statement or null if all root types have default values.
  */
@@ -52,11 +52,11 @@ function getRootSchemaStatement({ rootTypeNames, rootTypeStatements, containerPr
 }
 
 /**
- * Gets the root type names.
- * Return the default root type names or the custom root type names if they are present in the container properties.
+ * Gets the root type names. Return the default root type names or the custom root type names if they are present in the
+ * container properties.
  *
- * @param {Object} param0
- * @param {ContainerDetails|undefined} param0.containerProperties - The container properties.
+ * @param {object} param0
+ * @param {ContainerDetails | undefined} param0.containerProperties - The container properties.
  * @returns {RootTypeNamesParameter} - The root type names.
  */
 function getRootTypeNames({ containerProperties }) {
@@ -92,11 +92,10 @@ function getRootTypeNames({ containerProperties }) {
 }
 
 /**
- * Gets the root types.
- * Iterate over the entities and get the root types.
- * For each root type, get the nested statements composed of the fields of the entities with the operation type equal to the root type.
+ * Gets the root types. Iterate over the entities and get the root types. For each root type, get the nested statements
+ * composed of the fields of the entities with the operation type equal to the root type.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {EntityIdToJsonSchemaMap} param0.entityIdToJsonSchemaMap - The entities JSON schema.
  * @param {EntityIdToPropertiesMap} param0.entityIdToPropertiesMap - The entity properties.
  * @param {RootTypeNamesParameter} param0.rootTypeNames - The root type names.
@@ -134,18 +133,18 @@ function getRootTypes({ entityIdToJsonSchemaMap, entityIdToPropertiesMap, rootTy
 }
 
 /**
- * Gets the root type.
- * Iterate over the entities and get the root type.
- * For each root type, get the nested statements composed of the fields of the entities with the operation type equal to the root type.
- * If there are no entities with the operation type equal to the root type, return null.
+ * Gets the root type. Iterate over the entities and get the root type. For each root type, get the nested statements
+ * composed of the fields of the entities with the operation type equal to the root type. If there are no entities with
+ * the operation type equal to the root type, return null.
  *
- * @param {Object} param0
+ * @param {object} param0
  * @param {EntityIdToJsonSchemaMap} param0.entityIdToJsonSchemaMap - The entities JSON schema.
  * @param {EntityIdToPropertiesMap} param0.entityIdToPropertiesMap - The entity properties.
  * @param {string} param0.rootTypeName - The root type name.
  * @param {IdToNameMap} param0.definitionsIdToNameMap - The definitions id to name map.
  * @param {string} param0.rootType - The root type.
- * @returns {FEStatement | null} - The root type or null if there are no entities with the operation type equal to the root type.
+ * @returns {FEStatement | null} - The root type or null if there are no entities with the operation type equal to the
+ *   root type.
  */
 function getRootType({
 	entityIdToJsonSchemaMap,
