+++ forward_engineering/mappers/unions.js
@@ -1,5 +1,5 @@
 /**
- * @import { UnionDefinitions, FEStatement, Union } from "../../types/types"
+ * @import {UnionDefinitions, FEStatement, Union, UnionMemberType} from "../../shared/types/types"
  */
 
 const { getDefinitionNameFromReferencePath } = require('../helpers/referenceHelper');
@@ -9,9 +9,9 @@ const { getDirectivesUsageStatement } = require('./directiveUsageStatements');
 /**
  * Map the union member types to a string.
  *
- * @param {Object} args - The arguments
- * @param {UnionMemberType[]} unionMemberTypes - The union member types with all properties
- * @return {string}
+ * @param {object} params - The arguments
+ * @param {UnionMemberType[]} params.unionMemberTypes - The union member types with all properties
+ * @returns {string}
  */
 const getUnionMemberTypes = ({ unionMemberTypes }) => {
 	return unionMemberTypes
@@ -27,7 +27,7 @@ const getUnionMemberTypes = ({ unionMemberTypes }) => {
 /**
  * Maps a union to an FEStatement.
  *
- * @param {Object} args - The arguments
+ * @param {object} args - The arguments
  * @param {string} args.name - The name of the union.
  * @param {Union} args.union - The union object with all properties
  * @returns {FEStatement}
@@ -44,9 +44,9 @@ const mapUnion = ({ name, union }) => {
 /**
  * Maps the union types to an array of FEStatement.
  *
- * @param {Object} args - The arguments
+ * @param {object} args - The arguments
  * @param {UnionDefinitions} args.unions - The union types schema.
- * @return {FEStatement[]}
+ * @returns {FEStatement[]}
  */
 const getUnions = ({ unions }) => {
 	return Object.entries(unions).map(([name, union]) => mapUnion({ name, union }));
