+++ package.json
@@ -120,11 +120,13 @@
 		"eslint": "9.20.1",
 		"eslint-config-prettier": "10.0.1",
 		"eslint-plugin-import": "2.31.0",
+		"eslint-plugin-jsdoc": "50.6.8",
 		"eslint-plugin-prettier": "5.2.3",
 		"eslint-plugin-unused-imports": "4.1.4",
 		"globals": "15.15.0",
 		"lint-staged": "15.4.3",
 		"prettier": "3.5.0",
+		"prettier-plugin-jsdoc": "1.3.2",
 		"simple-git-hooks": "2.11.1"
 	}
 }
