+++ reverse_engineering/helpers/convertIntrospectionSchemaToGraphQLSchema.js
@@ -1,11 +1,12 @@
 /**
- * @import { IntrospectionQuery } from 'graphql';
+ * @import {IntrospectionQuery} from 'graphql';
  */
 
 const { buildClientSchema, printSchema } = require('graphql/utilities');
 
 /**
  * Convert introspection schema to GraphQL SDL schema
+ *
  * @param {IntrospectionQuery} introspectionSchema
  * @returns {string}
  */
