+++ reverse_engineering/helpers/sortByName.js
@@ -1,13 +1,14 @@
 /**
- * @import { FieldsOrder } from "../types/types"
+ * @import {FieldsOrder} from "../../shared/types/types"
  */
 
 /**
  * Sorts an array of objects by the name according to the fields order option
- * @param {Object} params
- * @param {Object[]} params.items - The items to sort
+ *
+ * @param {object} params
+ * @param {object[]} params.items - The items to sort
  * @param {FieldsOrder} params.fieldsOrder - The fields order
- * @returns {Object[]} The sorted items
+ * @returns {object[]} The sorted items
  */
 function sortByName({ items, fieldsOrder }) {
 	if (!Array.isArray(items)) {
