+++ reverse_engineering/mappers/typeDefinitions/typeDefinitions.js
@@ -1,18 +1,19 @@
+/**
+ * @import {DirectiveDefinition, FieldsOrder} from "../../../shared/types/types"
+ */
+
 const { astNodeKind } = require('../../constants/graphqlAST');
 const { findNodesByKind } = require('../../helpers/findNodesByKind');
 const { sortByName } = require('../../helpers/sortByName');
 const { getDirectiveTypeDefinitions } = require('./directive');
 
 /**
- * @import { DirectiveDefinition, FieldsOrder } from "../../types/types"
- */
-
-/**
  * Gets the type definitions structure
- * @param {Object} params
- * @param {Object[]} params.typeDefinitions - The type definitions nodes
+ *
+ * @param {object} params
+ * @param {object[]} params.typeDefinitions - The type definitions nodes
  * @param {FieldsOrder} params.fieldsOrder - The fields order
- * @returns {Object} The mapped type definitions
+ * @returns {object} The mapped type definitions
  */
 function getTypeDefinitions({ typeDefinitions, fieldsOrder }) {
 	const directives = getDirectiveTypeDefinitions({
@@ -26,10 +27,11 @@ function getTypeDefinitions({ typeDefinitions, fieldsOrder }) {
 
 /**
  * Creates the model definitions structure
- * @param {Object} params
+ *
+ * @param {object} params
  * @param {FieldsOrder} params.fieldsOrder - The fields order
  * @param {DirectiveDefinition[]} params.directives - The directive definitions
- * @returns {Object} The type definitions structure
+ * @returns {object} The type definitions structure
  */
 function getTypeDefinitionsStructure({ fieldsOrder, directives }) {
 	const definitions = {
@@ -47,11 +49,12 @@ function getTypeDefinitionsStructure({ fieldsOrder, directives }) {
 
 /**
  * Creates a definition category structure
- * @param {Object} params
+ *
+ * @param {object} params
  * @param {FieldsOrder} params.fieldsOrder - The fields order
  * @param {string} params.subtype - The subtype of the definition
- * @param {Object[]} params.properties - The properties to structure
- * @returns {Object} The definition category structure
+ * @param {object[]} params.properties - The properties to structure
+ * @returns {object} The definition category structure
  */
 function getDefinitionCategoryStructure({ fieldsOrder, subtype, properties }) {
 	const sortedFields = sortByName({ items: properties, fieldsOrder });
