+++ shared/types/fe.d.ts
@@ -1,3 +1,5 @@
+import { DirectiveLocations } from './shared';
+
 export type FEStatement = {
 	statement: string;
 	description?: string;
@@ -21,7 +23,7 @@ export type ObjectLikeTypeDefinition = {
 	implementsInterfaces?: ImplementsInterface[]; // Interfaces that the object type implements
 	typeDirectives?: DirectivePropertyData[]; // Directives for the type
 	properties: Record<string, FieldData>; // Properties of the object type
-}
+};
 
 // Field data type
 export type FieldData = RegularFieldData | ReferenceFieldData;
@@ -34,7 +36,7 @@ type RegularFieldData = {
 	items?: ArrayItem | ArrayItem[]; // Items of the List type
 	arguments?: Argument[]; // Arguments of the field
 	default?: unknown; // Default value of the field
-}
+};
 
 type ReferenceFieldData = {
 	$ref: string; // Reference path to the type definition
@@ -43,7 +45,7 @@ type ReferenceFieldData = {
 	fieldDirectives?: DirectivePropertyData[]; // Directives for the field
 	arguments?: Argument[]; // Arguments of the field
 	default?: unknown; // Default value of the reference
-}
+};
 
 export type ArrayItem = FieldData & {
 	required?: boolean; // If the array item is required
@@ -53,7 +55,7 @@ export type ArrayItem = FieldData & {
 type ArgumentListItem = {
 	type?: string;
 	required?: boolean;
-}
+};
 
 export type Argument = {
 	id: string;
@@ -66,25 +68,15 @@ export type Argument = {
 	listItems?: ArgumentListItem[];
 };
 
-// Directives
-export type DirectiveLocations = {
-	GUID: string;
-	argumentDefinition?: boolean;
-	enum?: boolean;
-	enumValue?: boolean;
-	field?: boolean;
-	fieldDefinition?: boolean;
-	inputFieldDefinition?: boolean;
-	inputObject?: boolean;
-	interface?: boolean;
-	mutation?: boolean;
-	object?: boolean;
-	query?: boolean;
-	scalar?: boolean;
-	schema?: boolean;
-	subscription?: boolean;
-	union?: boolean;
+export type ArgumentsResultStatement = {
+	argumentsStatement: string; // The formatted arguments string.
+	argumentsWarningComment: string; // The warning comment if any argument is missing a type.
 }
+
+// Directives
+export type FEDirectiveLocations = DirectiveLocations & {
+	GUID: string;
+};
 
 export type Directive = {
 	GUID: string;
@@ -95,11 +87,11 @@ export type Directive = {
 	ignore_z_value: boolean;
 	isActivated?: boolean;
 	schemaType: string;
-	directiveLocations?: DirectiveLocations;
+	directiveLocations?: FEDirectiveLocations;
 	arguments?: Argument[];
-}
+};
 
-export type DirectiveDefinitions = Record<string, Directive>
+export type DirectiveDefinitions = Record<string, Directive>;
 
 export type DirectivePropertyData = {
 	directiveFormat: 'Structured' | 'Raw'; // Format of the directive
@@ -125,13 +117,13 @@ type UnionMemberType = {
 	GUID: string;
 	displayName: string;
 	isActivated: boolean;
-}
+};
 
 type OneOfMeta = {
 	choice: string;
 	index: number;
 	isActivated: boolean;
-}
+};
 
 export type Union = {
 	type: 'union';
@@ -146,16 +138,16 @@ export type Union = {
 	oneOf_meta: OneOfMeta;
 	schemaType: string;
 	snippet: 'union';
-}
+};
 
-export type UnionDefinitions = Record<string, Union>
+export type UnionDefinitions = Record<string, Union>;
 
 // Root types
 export type RootTypeNamesParameter = {
 	query: string;
 	mutation: string;
 	subscription: string;
-}
+};
 
 export type EntityIdToJsonSchemaMap = Record<string, string>;
 export type EntityIdToPropertiesMap = Record<string, object[]>;
@@ -167,7 +159,7 @@ type ModelDetails = {
 	version: string;
 	dbVendor: string;
 	isLineageEnabled: boolean;
-}
+};
 
 type LineageSource = {
 	id: string;
@@ -176,11 +168,11 @@ type LineageSource = {
 	lineageSourceType: string;
 	lineageSourceFormat: string;
 	lineageFilePathName: string;
-}
+};
 
 type Lineage = {
 	sources: LineageSource[];
-}
+};
 
 export type ContainerDetails = {
 	name: string;
@@ -191,7 +183,7 @@ export type ContainerDetails = {
 	businessName?: string;
 	schemaRootTypes: RootTypeNamesParameter;
 	graphDirectives: DirectivePropertyData[];
-}
+};
 
 export type ContainerLevelScriptFEData = {
 	modelData: [ModelDetails, Lineage?]; // model level data
@@ -204,7 +196,7 @@ export type ContainerLevelScriptFEData = {
 	options: {
 		additionalOptions: object[]; // additional options
 		origin: string; // "ui" if the script is called from the forward engineering tab
-	}
+	};
 };
 
 export type GenerateContainerLevelScriptCallback = (error: Error | null, script?: string) => void;
@@ -214,6 +206,6 @@ export type ValidationResponseItem = {
 	label: string; // The label for the entity, typically indicating the location.
 	title: string; // The title of the entity, typically the error message.
 	context?: string; // The context of the entity, typically additional information.
-}
+};
 
 export type ValidateScriptCallback = (error: Error | null, validationErrors?: ValidationResponseItem[]) => void;
