+++ shared/types/types.d.ts
@@ -1,14 +1,3 @@
-type LogType = 'info' | 'error';
-
-type LogData = {
-	message?: string;
-	error?: Error;
-} | Error;
-
-type LogTitle = string;
-
-export type Logger = {
-	log: (logType: LogType, logData: LogData, logTitle: LogTitle, hiddenKeys: string[]) => void;
-	clear: () => void;
-	progress: (data: object) => void;
-};
+export type * from './shared.d.ts';
+export type * from './fe.d.ts';
+export type * from './re.d.ts';
