+++ types/List.json
@@ -16,38 +16,12 @@
 	},
 	"subtypes": {
 		"outputList": {
-			"childValueType": [
-				"String", 
-				"ID", 
-				"Int", 
-				"Float", 
-				"Boolean", 
-				"List", 
-				"reference"
-			],
-			"childReferenceValueType": [
-				"object",
-				"interface",
-				"union",
-				"enum",
-				"scalar"
-			]
+			"childValueType": ["String", "ID", "Int", "Float", "Boolean", "List", "reference"],
+			"childReferenceValueType": ["object", "interface", "union", "enum", "scalar"]
 		},
 		"inputList": {
-			"childValueType": [
-				"String", 
-				"ID", 
-				"Int", 
-				"Float", 
-				"Boolean", 
-				"List", 
-				"reference"
-			],
-			"childReferenceValueType": [
-				"input",
-				"enum",
-				"scalar"
-			]
+			"childValueType": ["String", "ID", "Int", "Float", "Boolean", "List", "reference"],
+			"childReferenceValueType": ["input", "enum", "scalar"]
 		}
 	}
 }
