+++ types/input.json
@@ -19,20 +19,8 @@
 	},
 	"subtypes": {
 		"snippetChildrenOnly": {
-			"childValueType": [
-				"String", 
-				"ID", 
-				"Int", 
-				"Float", 
-				"Boolean", 
-				"List", 
-				"reference"
-			],
-			"childReferenceValueType": [
-				"input",
-				"enum",
-				"scalar"
-			]
+			"childValueType": ["String", "ID", "Int", "Float", "Boolean", "List", "reference"],
+			"childReferenceValueType": ["input", "enum", "scalar"]
 		}
 	}
 }
