+++ types/interface.json
@@ -19,22 +19,8 @@
 	},
 	"subtypes": {
 		"snippetChildrenOnly": {
-			"childValueType": [
-				"String",
-				"ID",
-				"Int",
-				"Float",
-				"Boolean",
-				"List",
-				"reference"
-			],
-			"childReferenceValueType": [
-				"object",
-				"interface",
-				"union",
-				"enum",
-				"scalar"
-			]
+			"childValueType": ["String", "ID", "Int", "Float", "Boolean", "List", "reference"],
+			"childReferenceValueType": ["object", "interface", "union", "enum", "scalar"]
 		}
 	}
 }
