+++ types/union.json
@@ -20,12 +20,8 @@
 	},
 	"subtypes": {
 		"snippetChildrenOnly": {
-			"childValueType": [
-				"reference"
-			],
-			"childReferenceValueType": [
-				"object"
-			]
+			"childValueType": ["reference"],
+			"childReferenceValueType": ["object"]
 		}
 	}
 }
